[CLN] Move BASE_URL to config file
